Extract login module to LoginService
