+++ Assets/Scripts/Auth/Login.cs
@@ -1,22 +1,12 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.Text;
-using UnityEngine;
 using UnityEngine.Networking;
+using GachaController.Auth;
+using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 
-class LoginData
-{
-    public string expire_in;
-    public string access_token;
-    public string token_type;
-}
-
-public class AccessTokenKey
-{
- public static string accessTokenKey = "accessTokenKey";   
-}
 
 public class Login : MonoBehaviour
 {
@@ -25,10 +15,14 @@ public class Login : MonoBehaviour
     [SerializeField] private InputField passwordField;
     [SerializeField] private Button submitButton;
 
+
     // Start is called before the first frame update
     void Start()
     {
-        if (LoadAccessToken() ==
+        
+        
+    var loginService = new LoginService();
+        if (loginService.LoadAccessToken() ==
             "")
         {
             Debug.Log("not logged in");
@@ -52,38 +46,20 @@ public class Login : MonoBehaviour
     {
         var email = userIdField.text;
         var password = passwordField.text;
-        Debug.Log(email);
-        Debug.Log(password);
-        if (email != "" && password != "")
+        var loginService = new LoginService();
+        var coroutine = loginService.Login(email, password);
+        yield return StartCoroutine(coroutine);
+        var isSuccess = (bool) coroutine.Current;
+        if (isSuccess)
         {
-            var url = "http://localhost:8000/api/auth/login";
-            WWWForm form = new WWWForm();
-            form.AddField("email", email);
-            form.AddField("password", password);
-            var request = UnityWebRequest.Post(url, form);
-            yield return request.SendWebRequest();
-            // Debug.Log(request.downloadHandler.text);
-            var loginData = JsonUtility.FromJson<LoginData>(request.downloadHandler.text);
-            // Debug.Log(request.downloadHandler.text);
-            SaveAccessToken(loginData.access_token);
-            Debug.Log(loginData.access_token);
-            var accessToken = LoadAccessToken();
-            Debug.Log(accessToken);
-            SceneManager.LoadScene("LoginScene");
+            SceneManager.LoadScene("MyCharactersScene");
+        }
+        else
+        {
+            Debug.Log("Login is failed");
         }
     }
 
-
-    void SaveAccessToken(string accessToken)
-    {
-        PlayerPrefs.SetString(AccessTokenKey.accessTokenKey, accessToken);
-    }
-
-    string LoadAccessToken()
-    {
-        return
-            PlayerPrefs.GetString(AccessTokenKey.accessTokenKey);
-    }
 
     void LoadMainScene()
     {
