+++ Assets/Scripts/Auth/LoginService.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 53440d8483e542868a81211d1ed0d56a
+timeCreated: 1603862187
